+++ pcdsdevices/attenuator.py
@@ -11,7 +11,7 @@ from ophyd.pv_positioner import PVPositioner
 from ophyd.signal import Signal, EpicsSignal, EpicsSignalRO
 
 from .inout import InOutPositioner
-from .mv_interface import FltMvInterface
+from .interface import FltMvInterface
 
 logger = logging.getLogger(__name__)
 MAX_FILTERS = 12
@@ -34,6 +34,8 @@ class Filter(InOutPositioner):
     thickness = Cpt(EpicsSignal, ':THICK', kind='config')
     material = Cpt(EpicsSignal, ':MATERIAL', kind='config')
     stuck = Cpt(EpicsSignal, ':IS_STUCK', kind='omitted')
+
+    tab_component_names = True
 
 
 class FeeFilter(InOutPositioner):
@@ -81,6 +83,8 @@ class AttBase(FltMvInterface, PVPositioner):
     _icon = 'fa.barcode'
     # Subscription Types
     SUB_STATE = 'state'
+    # Tab complete whitelist
+    tab_whitelist = ['set_energy']
 
     def __init__(self, prefix, *, name, **kwargs):
         super().__init__(prefix, name=name, limits=(0, 1), **kwargs)