+++ pcdsdevices/state.py
@@ -7,11 +7,12 @@ from enum import Enum
 
 from ophyd.positioner import PositionerBase
 from ophyd.status import wait as status_wait, SubscriptionStatus
-from ophyd.signal import EpicsSignal, EpicsSignalRO
-from ophyd.device import Device, Component as Cpt, FormattedComponent as FCpt
+from ophyd.signal import EpicsSignal
+from ophyd.device import Device, Component as Cpt
 
 from .doc_stubs import basic_positioner_init
-from .mv_interface import MvInterface
+from .epics_motor import IMS
+from .interface import MvInterface
 from .signal import AggregateSignal
 
 logger = logging.getLogger(__name__)
@@ -392,12 +393,11 @@ class StateRecordPositioner(StatePositioner):
     settings, in the same order as the state enum. Unknown must be omitted.
     """
     state = Cpt(EpicsSignal, '', write_pv=':GO', kind='hinted')
-    readback = FCpt(EpicsSignalRO, '{self.prefix}:{self._readback}',
-                    kind='normal')
+    motor = Cpt(IMS, ':MOTOR', kind='normal')
+
+    tab_whitelist = ['motor']
 
     def __init__(self, prefix, *, name, **kwargs):
-        some_state = self.states_list[0]
-        self._readback = '{}_CALC.A'.format(some_state)
         super().__init__(prefix, name=name, **kwargs)
         self._has_subscribed_readback = False
         self._has_checked_state_enum = False
@@ -406,7 +406,8 @@ class StateRecordPositioner(StatePositioner):
         cid = super().subscribe(cb, event_type=event_type, run=run)
         if (event_type == self.SUB_READBACK and not
                 self._has_subscribed_readback):
-            self.readback.subscribe(self._run_sub_readback, run=False)
+            self.motor.user_readback.subscribe(self._run_sub_readback,
+                                               run=False)
             self._has_subscribed_readback = True
         return cid